msvc-auth/requirements.txt
flask==2.0.1
pika==1.2.0
sqlalchemy==1.4.23
pydantic==1.8.2
python-dotenv==0.19.0
gunicorn==21.2.0
werkzeug==2.0.3
psycopg2-binary==2.9.1
pyjwt
